Limit god address invites count (#271)

+++ common/math/integer.go
@@ -119,7 +119,7 @@ func MinInt(x, y int) int {
 	return y
 }
 
-func MaxInt(x, y uint64) uint64 {
+func MaxInt(x, y int) int {
 	if x > y {
 		return x
 	}
